moved the model definition stuff
